+++ R/pctopt.R
@@ -1,14 +1,14 @@
 #' Optimizing cut-off percentage for error minimization
 #'
 #' @description
-#' Finds an optimal cut-off percentage for error minimization (in vqssub, vqsassess, vqscustompct functions) that can decrease the number of singleton haplotypes to less than the desired percentage of the total reads.
+#' Finds an optimal cut-off percentage for error minimization (in vqssub, vqsassess, and vqscustompct functions) that can decrease the number of singleton haplotypes to less than the desired percentage of the total reads.
 #'
 #' @param fasta Input as a read alignment in FASTA format
 #' @param pctsing The desired percentage of singleton haplotypes relative to the total reads in the alignment.
 #' @param method Sequencing error minimization methods that replace low frequency nucleotide base (less than the "pct" cut-off) with consensus base of that position ("conbase": default) or with base of the dominant haplotype ("domhapbase").
 #' @param samplingfirst Downsampling before (TRUE) or after (FALSE: default) the error minimization.
 #' @param gappct The percent cut-off particularly specified for gap (-). If it is not specified or less than "pct", "gappct" will be equal to "pct" (default).
-#' @param ignoregappositions Replace all nucleotides in the positions in the alignment containing gap(s) with gap. This will make such positions no longer SNV. The default is "FALSE".
+#' @param ignoregappositions Replace all nucleotides in the positions in the alignment containing gap(s) with gap. This will make such positions no longer single nucleotide variant (SNV). The default is "FALSE".
 #' @param samsize Sample size (number of reads) after down-sampling. If it is not specified or more than number of reads in the original alignment, down-sampling will not be performed (default).
 #' @param label String within quotation marks indicating name of read alignment (optional).
 #'
@@ -121,7 +121,7 @@ pctopt <- function(fasta, pctsing = 0, method= c("conbase", "domhapbase"), sampl
       maxfrq$position <- as.integer(maxfrq$position)
       maxfrq$maxfrq = toupper(maxfrq$maxfrq)
       if(dim(lowfrq)[1] == 0){
-        print("No low frequency snv")
+        message("No low frequency SNV")
       }else{
         lowfrq <- merge(x = lowfrq, y =  maxfrq, by = "position", all.x = TRUE)
       }
@@ -131,7 +131,7 @@ pctopt <- function(fasta, pctsing = 0, method= c("conbase", "domhapbase"), sampl
 
       for (i in 1:nrow(lowfrq)){
         if(dim(lowfrq)[1] == 0){
-          print("No low frequency snv")
+          message("No low frequency SNV")
         }else{
           dfseq[,lowfrq[i,1]][dfseq[,lowfrq[i,1]] == lowfrq[i,2]] <- lowfrq[i,3]
         }
@@ -197,7 +197,7 @@ pctopt <- function(fasta, pctsing = 0, method= c("conbase", "domhapbase"), sampl
       }
     }
     if(missing(label)){
-      print(pct)
+      return(pct)
     }else{
       return(store)
     }
@@ -263,7 +263,6 @@ pctopt <- function(fasta, pctsing = 0, method= c("conbase", "domhapbase"), sampl
       if(domhap$nr[1] == 1){
         stop("No dominant haplotype, please use conbase method instead")
       }else{
-        print(paste0("Dominant haplotype is ", format(round(domhap$nr[1]*100/sum(domhap$nr), 2), nsmall = 2), "% of total reads"))
         domhap <- dss2df(domhap$hseqs)
       }
 
@@ -275,7 +274,7 @@ pctopt <- function(fasta, pctsing = 0, method= c("conbase", "domhapbase"), sampl
       colnames(domhap) <- c("domhap", "position")
 
       if(dim(lowfrq)[1] == 0){
-        print("No low frequency snv")
+        message("No low frequency SNV")
       }else{
         lowfrq <- merge(x = lowfrq, y =  domhap, by = "position", all.x = TRUE)
       }
@@ -285,7 +284,7 @@ pctopt <- function(fasta, pctsing = 0, method= c("conbase", "domhapbase"), sampl
 
       for (i in 1:nrow(lowfrq)){
         if(dim(lowfrq)[1] == 0){
-          print("No low frequency snv")
+          message("No low frequency SNV")
         }else{
           dfseq[,lowfrq[i,1]][dfseq[,lowfrq[i,1]] == lowfrq[i,2]] <- lowfrq[i,3]
         }
@@ -351,7 +350,7 @@ pctopt <- function(fasta, pctsing = 0, method= c("conbase", "domhapbase"), sampl
       }
     }
     if(missing(label)){
-      print(pct)
+      return(pct)
     }else{
       return(store)
     }