+++ R/vqssub.R
@@ -8,7 +8,7 @@
 #' @param samplingfirst Downsampling before (TRUE) or after (FALSE: default) the error minimization.
 #' @param pct Percent cut-off defining low frequency nucleotide base that will be replaced (must be specified).
 #' @param gappct The percent cut-off particularly specified for gap (-). If it is not specified or less than "pct", "gappct" will be equal to "pct" (default).
-#' @param ignoregappositions Replace all nucleotides in the positions in the alignment containing gap(s) with gap. This will make such positions no longer SNV. The default is "FALSE".
+#' @param ignoregappositions Replace all nucleotides in the positions in the alignment containing gap(s) with gap. This will make such positions no longer single nucleotide variant (SNV). The default is "FALSE".
 #' @param samsize Sample size (number of reads) after down-sampling. If it is not specified or more than number of reads in the original alignment, down-sampling will not be performed (default).
 #' @param label String within quotation marks indicating name of read alignment (optional).
 #'
@@ -98,7 +98,7 @@ vqssub <- function(fasta, method= c("conbase", "domhapbase"), samplingfirst = FA
     frqpc <- as.data.frame(t(frqpc))
 
     if(missing(pct)){
-      stop("Please specify % cut-off for low SNV frequency (pct)")
+      stop("Please specify % cut-off for low frequency SNV (pct)")
     }else{
       lowfrq <- apply(frqpc, 1, function(x) paste(names(which(x[1:5] < pct & x[1:5] > 0)), collapse = ","))
     }
@@ -132,7 +132,7 @@ vqssub <- function(fasta, method= c("conbase", "domhapbase"), samplingfirst = FA
     maxfrq$position <- as.integer(maxfrq$position)
     maxfrq$maxfrq = toupper(maxfrq$maxfrq)
     if(dim(lowfrq)[1] == 0){
-      print("No low frequency snv")
+      message("No low frequency SNV")
     }else{
       lowfrq <- merge(x = lowfrq, y =  maxfrq, by = "position", all.x = TRUE)
     }
@@ -142,7 +142,7 @@ vqssub <- function(fasta, method= c("conbase", "domhapbase"), samplingfirst = FA
 
     for (i in 1:nrow(lowfrq)){
       if(dim(lowfrq)[1] == 0){
-        print("No low frequency snv")
+        message("No low frequency SNV")
       }else{
         dfseq[,lowfrq[i,1]][dfseq[,lowfrq[i,1]] == lowfrq[i,2]] <- lowfrq[i,3]
       }
@@ -260,7 +260,7 @@ vqssub <- function(fasta, method= c("conbase", "domhapbase"), samplingfirst = FA
     frqpc <- as.data.frame(t(frqpc))
 
     if(missing(pct)){
-      stop("Please specify % cut-off for low SNV frequency (pct)")
+      stop("Please specify % cut-off for low frequency SNV (pct)")
     }else{
       lowfrq <- apply(frqpc, 1, function(x) paste(names(which(x[1:5] < pct & x[1:5] > 0)), collapse = ","))
     }
@@ -290,7 +290,7 @@ vqssub <- function(fasta, method= c("conbase", "domhapbase"), samplingfirst = FA
     if(domhap$nr[1] == 1){
       stop("No dominant haplotype, please use conbase method instead")
     }else{
-      print(paste0("Dominant haplotype is ", format(round(domhap$nr[1]*100/sum(domhap$nr), 2), nsmall = 2), "% of total reads"))
+      message(paste0("Dominant haplotype is ", format(round(domhap$nr[1]*100/sum(domhap$nr), 2), nsmall = 2), "% of total reads"))
       domhap <- dss2df(domhap$hseqs)
     }
 
@@ -302,7 +302,7 @@ vqssub <- function(fasta, method= c("conbase", "domhapbase"), samplingfirst = FA
     colnames(domhap) <- c("domhap", "position")
 
     if(dim(lowfrq)[1] == 0){
-      print("No low frequency snv")
+      message("No low frequency SNV")
     }else{
       lowfrq <- merge(x = lowfrq, y =  domhap, by = "position", all.x = TRUE)
     }
@@ -312,7 +312,7 @@ vqssub <- function(fasta, method= c("conbase", "domhapbase"), samplingfirst = FA
 
     for (i in 1:nrow(lowfrq)){
       if(dim(lowfrq)[1] == 0){
-        print("No low frequency snv")
+        message("No low frequency SNV")
       }else{
         dfseq[,lowfrq[i,1]][dfseq[,lowfrq[i,1]] == lowfrq[i,2]] <- lowfrq[i,3]
       }